pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inkwell"
version = "0.1.0"
description = "Self-hosted, BYOK outreach intelligence. Scan forums free; only draft replies in your voice spend tokens."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Saurabh Singh"},
]
keywords = ["outreach", "reddit", "ai", "byok", "self-hosted", "persona", "writing", "open-source"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "python-dotenv>=1.0.0",
    "openai>=1.30.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "litellm>=1.40.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[project.scripts]
inkwell = "inkwell.__main__:main"

[tool.setuptools.packages.find]
include = ["inkwell*"]

[tool.ruff]
line-length = 100
target-version = "py311"
